// ui/src/generated/BrowserCommand.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { BrowserQuery } from "./BrowserQuery";

export type BrowserCommand = { "type": "Query", query: BrowserQuery, } | { "type": "ListRoots" } | { "type": "SetFavourite", item: string, favourite: boolean, } | { "type": "SetTags", item: string, tags: Array<string>, } | { "type": "AddFolder", path: string, } | { "type": "RemoveFolder", root: string, } | { "type": "Rescan", root: string | null, } | { "type": "Preview", item: string, sync: boolean, };